REFACTOR - started writing tests
